test: use afterAll hook assertion instead of cli
