+++ tests/basic/test/meta.test.ts
@@ -7,6 +7,7 @@ describe('Meta', () => {
     expect(__filename.endsWith('/basic/test/meta.test.ts')).toBeTruthy();
   });
 
+  // TODO
   it.todo('should get test file import meta correctly', async () => {
     expect(import.meta.dirname.includes('/basic/test')).toBeTruthy();
     expect(
@@ -14,6 +15,7 @@ describe('Meta', () => {
     ).toBeTruthy();
   });
 
+  // TODO
   it.todo('should get source file meta correctly', async () => {
     expect(aDirName.endsWith('/basic/src')).toBeTruthy();
     expect(aFileName.endsWith('/basic/src/meta.ts')).toBeTruthy();
