+++ tests/filter/testNamePattern.test.ts
@@ -7,32 +7,6 @@ const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
 describe('test testNamePattern', () => {
-  it('should run success without filter', async () => {
-    const { cli } = await runRstestCli({
-      command: 'rstest',
-      args: ['run', 'fixtures/testNamePattern.test.ts'],
-      options: {
-        nodeOptions: {
-          cwd: __dirname,
-        },
-      },
-    });
-
-    await cli.exec;
-    expect(cli.exec.process?.exitCode).toBe(0);
-
-    const logs = cli.stdout.split('\n').filter(Boolean);
-
-    expect(logs.filter((log) => log.startsWith('['))).toMatchInlineSnapshot(`
-      [
-        "[test] in level-B-A",
-        "[test] in level-B-C-A",
-        "[test] in level-C",
-        "[test] in level-D-A",
-      ]
-    `);
-  });
-
   it('should filter test case name success', async () => {
     const { cli } = await runRstestCli({
       command: 'rstest',
@@ -60,34 +34,6 @@ describe('test testNamePattern', () => {
     const { cli } = await runRstestCli({
       command: 'rstest',
       args: ['run', 'fixtures/testNamePattern.test.ts', '-t=level-B'],
-      options: {
-        nodeOptions: {
-          cwd: __dirname,
-        },
-      },
-    });
-
-    await cli.exec;
-    expect(cli.exec.process?.exitCode).toBe(0);
-
-    const logs = cli.stdout.split('\n').filter(Boolean);
-
-    expect(logs.filter((log) => log.startsWith('['))).toMatchInlineSnapshot(`
-      [
-        "[test] in level-B-A",
-        "[test] in level-B-C-A",
-      ]
-    `);
-  });
-
-  it('should filter test suite name success use regex', async () => {
-    const { cli } = await runRstestCli({
-      command: 'rstest',
-      args: [
-        'run',
-        'fixtures/testNamePattern.test.ts',
-        '-c=fixtures/testNamePattern.config.ts',
-      ],
       options: {
         nodeOptions: {
           cwd: __dirname,
