+++ tests/runner/test/async.test.ts
@@ -1,40 +1,79 @@
-import { dirname } from 'node:path';
-import { fileURLToPath } from 'node:url';
-import { describe, expect, it } from '@rstest/core';
-import { getTestName, runRstestCli } from '../../scripts';
+import { afterAll, describe, expect, it } from '@rstest/core';
+import { sleep } from '../../scripts/utils';
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
+const logs: string[] = [];
 
-describe('Test Async Suite', () => {
-  it('should run async suite in the correct order', async () => {
-    const { cli } = await runRstestCli({
-      command: 'rstest',
-      args: ['run', 'fixtures/async.test.ts'],
-      options: {
-        nodeOptions: {
-          cwd: __dirname,
-        },
-      },
+afterAll(() => {
+  expect(logs).toEqual([
+    'run 0-0',
+    'run 0-1-0',
+    'run 0-1-1-0',
+    'run 0-2-0',
+    'run 0-3',
+    'run 1-0',
+    'run 2-0-0',
+    'run 2-1',
+    'run 3-0',
+  ]);
+});
+
+describe('should run async suite in the correct order', () => {
+  describe('0', async () => {
+    await sleep(100);
+
+    it('0-0', () => {
+      logs.push('run 0-0');
+      expect(1 + 1).toBe(2);
     });
 
-    await cli.exec;
+    describe('0-1', async () => {
+      it('0-1-0', () => {
+        logs.push('run 0-1-0');
+        expect(1 + 1).toBe(2);
+      });
 
-    const logs = cli.stdout.split('\n').filter(Boolean);
+      describe('0-1-1', () => {
+        it('0-1-1-0', () => {
+          logs.push('run 0-1-1-0');
+          expect(1 + 1).toBe(2);
+        });
+      });
+    });
 
-    // test execution order
-    expect(logs.filter((log) => log.startsWith('run'))).toMatchInlineSnapshot(`
-      [
-        "run 0-0",
-        "run 0-1-0",
-        "run 0-1-1-0",
-        "run 0-2-0",
-        "run 0-3",
-        "run 1-0",
-        "run 2-0-0",
-        "run 2-1",
-        "run 3-0",
-      ]
-    `);
+    describe('0-2', () => {
+      it('0-2-0', () => {
+        logs.push('run 0-2-0');
+        expect(1 + 1).toBe(2);
+      });
+    });
+
+    it('0-3', () => {
+      logs.push('run 0-3');
+      expect(1 + 1).toBe(2);
+    });
+  });
+
+  it('1', () => {
+    logs.push('run 1-0');
+    expect(1 + 1).toBe(2);
+  });
+
+  describe('2', () => {
+    describe('2-0', async () => {
+      it('2-0-0', () => {
+        logs.push('run 2-0-0');
+        expect(1 + 1).toBe(2);
+      });
+    });
+
+    it('2-1', () => {
+      logs.push('run 2-1');
+      expect(1 + 1).toBe(2);
+    });
+  });
+
+  it('3', () => {
+    logs.push('run 3-0');
+    expect(1 + 1).toBe(2);
   });
 });
