+++ tests/spy/withImplementation.test.ts
@@ -1,43 +1,93 @@
-import { dirname } from 'node:path';
-import { fileURLToPath } from 'node:url';
-import { expect, it } from '@rstest/core';
-import { runRstestCli } from '../scripts';
+import { afterAll, describe, expect, it, rstest } from '@rstest/core';
+const logs: string[] = [];
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
+afterAll(() => {
+  expect(logs).toEqual([
+    '[call callback]',
+    '[call temp]',
+    '[call temp]',
+    '[callback res] temp temp',
+    '[call myMockFn]',
+    '[call original - 1]',
+    '[call myMockFn - 1]',
+    '[call original]',
+    '[1 - call callback]',
+    '[1 - call temp]',
+    '[1 - callback res] temp',
+    '[1 - call myMockFn]',
+    '[1 - call original]',
+    '[1 - call myMockFn - 1]',
+    '[1 - call original]',
+  ]);
+});
 
-it('test withImplementation', async () => {
-  const { cli } = await runRstestCli({
-    command: 'rstest',
-    args: ['run', 'fixtures/withImplementation.test'],
-    options: {
-      nodeOptions: {
-        cwd: __dirname,
+describe('test withImplementation', () => {
+  it('withImplementation', () => {
+    let isMockCalled = false;
+    const mockFn = () => {
+      isMockCalled = true;
+      logs.push('[call original]');
+      return 'original';
+    };
+    const myMockFn = rstest.fn(mockFn);
+
+    const mockFn1 = () => {
+      isMockCalled = true;
+      logs.push('[call original - 1]');
+      return 'original - 1';
+    };
+
+    myMockFn.mockImplementationOnce(mockFn1);
+
+    myMockFn.withImplementation(
+      () => {
+        logs.push('[call temp]');
+        return 'temp';
       },
-    },
+      () => {
+        logs.push('[call callback]');
+        const res = myMockFn();
+        const res1 = myMockFn();
+        logs.push(`[callback res] ${res} ${res1}`);
+      },
+    );
+
+    expect(myMockFn.getMockImplementation()).toBe(mockFn1);
+    expect(isMockCalled).toBe(false);
+
+    logs.push('[call myMockFn]');
+    expect(myMockFn()).toBe('original - 1');
+    expect(isMockCalled).toBe(true);
+
+    logs.push('[call myMockFn - 1]');
+
+    expect(myMockFn()).toBe('original');
+    expect(myMockFn).toHaveBeenCalledTimes(4);
   });
 
-  await cli.exec;
-  expect(cli.exec.process?.exitCode).toBe(0);
-  const logs = cli.stdout.split('\n').filter(Boolean);
+  it('withImplementation async', async () => {
+    const myMockFn = rstest.fn(() => {
+      logs.push('[1 - call original]');
+      return 'original';
+    });
 
-  expect(logs.filter((log) => log.startsWith('['))).toMatchInlineSnapshot(`
-    [
-      "[call callback]",
-      "[call temp]",
-      "[call temp]",
-      "[callback res] temp temp",
-      "[call myMockFn]",
-      "[call original - 1]",
-      "[call myMockFn - 1]",
-      "[call original]",
-      "[1 - call callback]",
-      "[1 - call temp]",
-      "[1 - callback res] temp",
-      "[1 - call myMockFn]",
-      "[1 - call original]",
-      "[1 - call myMockFn - 1]",
-      "[1 - call original]",
-    ]
-  `);
+    await myMockFn.withImplementation(
+      () => {
+        logs.push('[1 - call temp]');
+        return 'temp';
+      },
+      async () => {
+        logs.push('[1 - call callback]');
+        const res = myMockFn();
+        logs.push(`[1 - callback res] ${res}`);
+      },
+    );
+
+    logs.push('[1 - call myMockFn]');
+
+    expect(myMockFn()).toBe('original');
+
+    logs.push('[1 - call myMockFn - 1]');
+    expect(myMockFn()).toBe('original');
+  });
 });
