+++ tests/test-api/chain.test.ts
@@ -38,42 +38,4 @@ describe('Test Chain', () => {
       ]
     `);
   });
-
-  it('Support only.each', async () => {
-    const { cli } = await runRstestCli({
-      command: 'rstest',
-      args: ['run', 'fixtures/only.each.test.ts'],
-      options: {
-        nodeOptions: {
-          cwd: __dirname,
-        },
-      },
-    });
-    await cli.exec;
-    expect(cli.exec.process?.exitCode).toBe(0);
-
-    const logs = cli.stdout.split('\n').filter(Boolean);
-    expect(
-      logs.find((log) => log.includes('Tests 3 passed | 1 skipped')),
-    ).toBeTruthy();
-  });
-
-  it('Support only.fails', async () => {
-    const { cli } = await runRstestCli({
-      command: 'rstest',
-      args: ['run', 'fixtures/only.fails.test.ts'],
-      options: {
-        nodeOptions: {
-          cwd: __dirname,
-        },
-      },
-    });
-    await cli.exec;
-    expect(cli.exec.process?.exitCode).toBe(0);
-
-    const logs = cli.stdout.split('\n').filter(Boolean);
-    expect(
-      logs.find((log) => log.includes('Tests 1 passed | 1 skipped')),
-    ).toBeTruthy();
-  });
 });
