--- tests/test-api/fixtures/only.fails.test.ts
@@ -1,9 +1,0 @@
-import { expect, it } from '@rstest/core';
-
-it.only.fails('will pass when failed', () => {
-  expect(1 + 1).toBe(1);
-});
-
-it('will not run', () => {
-  expect(1 + 1).toBe(1);
-});
