--- tests/test-api/fixtures/sequential.test.ts
@@ -1,32 +1,0 @@
-import { describe, it } from '@rstest/core';
-
-describe.concurrent('suite', () => {
-  it('test', async () => {
-    console.log('[log] test');
-    await new Promise((resolve) => setTimeout(resolve, 10));
-    console.log('[log] test 0 - 1');
-  });
-  it('test 1', async () => {
-    console.log('[log] test 1');
-    await new Promise((resolve) => setTimeout(resolve, 200));
-    console.log('[log] test 1 - 1');
-  });
-
-  it.sequential('test 2', async () => {
-    console.log('[log] test 2');
-    await new Promise((resolve) => setTimeout(resolve, 100));
-    console.log('[log] test 2 - 1');
-  });
-
-  it('test 3', async () => {
-    console.log('[log] test 3');
-    await new Promise((resolve) => setTimeout(resolve, 100));
-    console.log('[log] test 3 - 1');
-  });
-
-  it('test 4', async () => {
-    console.log('[log] test 4');
-    await new Promise((resolve) => setTimeout(resolve, 100));
-    console.log('[log] test 4 - 1');
-  });
-});
