+++ tests/test-api/index.test.ts
@@ -31,41 +31,4 @@ describe('Test API', () => {
       ),
     ).toBeTruthy();
   });
-
-  it('test only', async () => {
-    const __filename = fileURLToPath(import.meta.url);
-    const __dirname = dirname(__filename);
-
-    const { cli } = await runRstestCli({
-      command: 'rstest',
-      args: ['run', 'fixtures/only.test.ts'],
-      options: {
-        nodeOptions: {
-          cwd: __dirname,
-        },
-      },
-    });
-    await cli.exec;
-    expect(cli.exec.process?.exitCode).toBe(0);
-
-    const logs = cli.stdout.split('\n').filter(Boolean);
-
-    expect(logs.filter((log) => log.startsWith('['))).toMatchInlineSnapshot(`
-          [
-            "[beforeEach] root",
-            "[test] in level A",
-            "[beforeEach] root",
-            "[test] in level B-B",
-            "[beforeEach] root",
-            "[test] in level D",
-          ]
-        `);
-
-    expect(
-      logs.find((log) => log.includes('Test Files 1 passed')),
-    ).toBeTruthy();
-    expect(
-      logs.find((log) => log.includes('Tests 3 passed | 3 skipped')),
-    ).toBeTruthy();
-  });
 });
