+++ tests/test-api/sequential.test.ts
@@ -1,40 +1,48 @@
-import { dirname } from 'node:path';
-import { fileURLToPath } from 'node:url';
-import { describe, expect, it } from '@rstest/core';
-import { runRstestCli } from '../scripts';
+import { afterAll, describe, expect, it } from '@rstest/core';
+const logs: string[] = [];
 
-describe('Test sequential', () => {
-  it('should run sequential cases correctly in concurrent suite', async () => {
-    const __filename = fileURLToPath(import.meta.url);
-    const __dirname = dirname(__filename);
+afterAll(() => {
+  expect(logs).toEqual([
+    '[log] test',
+    '[log] test 1',
+    '[log] test 0 - 1',
+    '[log] test 1 - 1',
+    '[log] test 2',
+    '[log] test 2 - 1',
+    '[log] test 3',
+    '[log] test 4',
+    '[log] test 3 - 1',
+    '[log] test 4 - 1',
+  ]);
+});
 
-    const { cli } = await runRstestCli({
-      command: 'rstest',
-      args: ['run', 'fixtures/sequential.test.ts'],
-      options: {
-        nodeOptions: {
-          cwd: __dirname,
-        },
-      },
-    });
-    await cli.exec;
-    expect(cli.exec.process?.exitCode).toBe(0);
+describe.concurrent('suite', () => {
+  it('test', async () => {
+    logs.push('[log] test');
+    await new Promise((resolve) => setTimeout(resolve, 10));
+    logs.push('[log] test 0 - 1');
+  });
+  it('test 1', async () => {
+    logs.push('[log] test 1');
+    await new Promise((resolve) => setTimeout(resolve, 200));
+    logs.push('[log] test 1 - 1');
+  });
 
-    const logs = cli.stdout.split('\n').filter(Boolean);
+  it.sequential('test 2', async () => {
+    logs.push('[log] test 2');
+    await new Promise((resolve) => setTimeout(resolve, 100));
+    logs.push('[log] test 2 - 1');
+  });
 
-    expect(logs.filter((log) => log.includes('[log]'))).toMatchInlineSnapshot(`
-      [
-        "[log] test",
-        "[log] test 1",
-        "[log] test 0 - 1",
-        "[log] test 1 - 1",
-        "[log] test 2",
-        "[log] test 2 - 1",
-        "[log] test 3",
-        "[log] test 4",
-        "[log] test 3 - 1",
-        "[log] test 4 - 1",
-      ]
-    `);
+  it('test 3', async () => {
+    logs.push('[log] test 3');
+    await new Promise((resolve) => setTimeout(resolve, 100));
+    logs.push('[log] test 3 - 1');
+  });
+
+  it('test 4', async () => {
+    logs.push('[log] test 4');
+    await new Promise((resolve) => setTimeout(resolve, 100));
+    logs.push('[log] test 4 - 1');
   });
 });
